Address runtime lifecycle PR followups
