+++ server/src/services/recovery/pause-hold-guard.ts
@@ -1,11 +1,14 @@
 import type { Db } from "@paperclipai/db";
 import { issueTreeControlService } from "../issue-tree-control.js";
 
+type IssueTreeControlService = ReturnType<typeof issueTreeControlService>;
+
 export async function isAutomaticRecoverySuppressedByPauseHold(
   db: Db,
   companyId: string,
   issueId: string,
+  treeControlSvc: IssueTreeControlService = issueTreeControlService(db),
 ) {
-  const activePauseHold = await issueTreeControlService(db).getActivePauseHoldGate(companyId, issueId);
+  const activePauseHold = await treeControlSvc.getActivePauseHoldGate(companyId, issueId);
   return Boolean(activePauseHold);
 }
